fix(pacmak): runtime dependency should match pacmak release
The code generated by `jsii-pacmak` has a dependency on jsii runtime
versions that are expected to match that of `jsii-pacmak` itself.
This fixes the .NET and Python generators using the assembly's
`jsiiVersion` property, which is a metadata field supposed to only
indicate what version of the compiler was used to build the assembly.